fix project edition feature

+++ src/main/java/com/resume/web/controller/ProjectController.java
@@ -1,18 +1,10 @@
 package com.resume.web.controller;
 
-import com.resume.converter.ConverterHelper;
+import com.resume.Services.IProjectService;
 import com.resume.dco.ProjectDco;
 import com.resume.dto.ProjectDto;
-import com.resume.model.Competence;
-import com.resume.model.Project;
-import com.resume.model.enums.ProjectTypeEnum;
-import com.resume.repository.CompetenceRepository;
-import com.resume.repository.ExperienceRepository;
-import com.resume.repository.ProjectRepository;
 import com.resume.web.exceptions.NoContentException;
-import com.resume.web.exceptions.UnexpectedCompetenceException;
-import com.resume.web.exceptions.UnexpectedExperienceException;
-import com.resume.web.serviceImpl.ProjectServiceImpl;
+import com.resume.web.exceptions.UnexpectedProjectException;
 import io.swagger.annotations.ApiOperation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,27 +13,16 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 @RestController
 public class ProjectController {
 
-    private ProjectServiceImpl projectService;
-
-    private ExperienceRepository experienceRepository;
-
-    private CompetenceRepository competenceRepository;
-
-    private ProjectRepository projectRepository;
+    private final IProjectService projectService;
 
     public static final Logger logger = LoggerFactory.getLogger(ProjectController.class);
 
-    public ProjectController(ProjectServiceImpl projectService, ExperienceRepository experienceRepository, ProjectRepository projectRepository, CompetenceRepository competenceRepository) {
+    public ProjectController(IProjectService projectService) {
         this.projectService = projectService;
-        this.experienceRepository = experienceRepository;
-        this.projectRepository = projectRepository;
-        this.competenceRepository = competenceRepository;
     }
 
 
@@ -68,7 +49,7 @@ public class ProjectController {
     @ApiOperation(value = "Post a project")
     @PostMapping(value = "/Projects")
     public ResponseEntity<ProjectDto> post(@Valid @RequestBody ProjectDco projectDco) {
-        return ResponseEntity.ok().body(projectService.postProject(ConverterHelper.convertToEntity(projectDco, experienceRepository, competenceRepository)));
+        return ResponseEntity.ok().body(projectService.postProject(projectDco));
     }
 
     @CrossOrigin()
@@ -77,39 +58,31 @@ public class ProjectController {
     public ResponseEntity<ProjectDto> put(@Valid @RequestBody ProjectDco projectDco) {
         logger.info("Updating Project with id {}", projectDco.getProjectId());
 
-        Project currentProject = projectRepository.findByProjectId(projectDco.getProjectId());
+        ProjectDto currentProject = projectService.getProject(projectDco.getProjectId());
 
         if (currentProject == null) {
             logger.error("Unable to update. Project with id {} not found.", projectDco.getProjectId());
-            throw new UnexpectedExperienceException("Unable to update. Project with id " + projectDco.getProjectId() + " not found.");
+            throw new UnexpectedProjectException("Unable to update. Project with id " + projectDco.getProjectId() + " not found.");
         }
 
-        currentProject.setDescription(projectDco.getDescription());
-        currentProject.setLinkToProject(projectDco.getLinkToProject());
-        currentProject.setProjectType(ProjectTypeEnum.valueOf(projectDco.getProjectType()));
-        currentProject.setTitle(projectDco.getTitle());
-        if(projectDco.getExperienceId() != null) {
-            currentProject.setExperience(experienceRepository.findByExperienceId(projectDco.getExperienceId()));
-        }
-        if(projectDco.getCompetenceIds() != null) {
-            Set<Competence> competencesForProject = projectDco.getCompetenceIds().stream().map(id -> competenceRepository.findByCompetenceId(id)).collect(Collectors.toSet());
-            currentProject.setCompetencesForProject(competencesForProject);
-        }
-        return ResponseEntity.ok().body(projectService.putProject(currentProject));
+        return ResponseEntity.ok().body(projectService.putProject(projectDco));
     }
 
     @CrossOrigin()
     @ApiOperation(value = "Delete a project")
     @DeleteMapping(value = "/Projects/{projectId}")
-    public ResponseEntity<Project> delete(@PathVariable Integer projectId) {
+    public ResponseEntity<ProjectDto> delete(@PathVariable Integer projectId) {
         logger.info("Deleting project with id {}", projectId);
-        Project projectToDelete = projectRepository.findByProjectId(projectId);
-        Integer deletedCompetence  = projectService.deleteProject(projectId);
+        ProjectDto projectToDelete = projectService.getProject(projectId);
 
-        if(deletedCompetence == 0) {
-            throw new UnexpectedCompetenceException("Unable to delete. Project with id " + projectId + " not found.");
+        if(projectToDelete != null) {
+            Integer deletedProjectId = projectService.deleteProject(projectId);
+            if (deletedProjectId == 0) {
+                throw new UnexpectedProjectException("Unable to delete. Project with id " + projectId + " not found.");
+            }
+        } else {
+            throw new UnexpectedProjectException("Unable to delete. Project with id " + projectId + " not found.");
         }
-        System.out.println(projectToDelete);
         return ResponseEntity.ok().body(projectToDelete);
     }
